backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-SocketIO==5.1.1
Flask-JWT-Extended==4.3.1
python-dotenv==0.19.0
SQLAlchemy==1.4.23
psycopg2-binary==2.9.1
gunicorn==20.1.0
eventlet==0.33.0
pandas==1.3.3
scikit-learn==0.24.2
numpy==1.21.2
pytest==6.2.5
python-socketio==5.4.0 